requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dateparser==1.1.8
fastapi==0.95.2
frozenlist==1.3.3
h11==0.14.0
idna==3.4
multidict==6.0.4
pycryptodome==3.18.0
pydantic==1.10.8
python-binance==1.0.17
python-dateutil==2.8.2
pytz==2023.3
regex==2023.5.5
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.6.2
tzdata==2023.3
tzlocal==5.0.1
ujson==5.7.0
urllib3==2.0.2
uvicorn==0.22.0
websockets==11.0.3
yarl==1.9.2
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dateparser==1.1.8
exceptiongroup==1.1.1
fastapi==0.95.2
frozenlist==1.3.3
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
multidict==6.0.4
packaging==23.1
pluggy==1.0.0
pycryptodome==3.18.0
pydantic==1.10.8
pytest==7.3.1
pytest-asyncio==0.21.0
python-binance==1.0.17
python-dateutil==2.8.2
pytz==2023.3
regex==2023.5.5
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.6.2
tzdata==2023.3
tzlocal==5.0.1
ujson==5.7.0
urllib3==2.0.2
uvicorn==0.22.0
websockets==11.0.3
yarl==1.9.2
